app/requirements.txt
annotated-types==0.7.0 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
asgiref==3.8.1 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
django-crispy-forms==1.14.0 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
django-extensions==3.2.3 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
django-qr-code==3.1.1 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
django-tables2==2.7.5 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
django==4.2.19 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
gunicorn==20.1.0 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
numpy==2.2.3 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
packaging==24.2 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
pillow==11.1.0 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
plotly==5.24.1 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
pydantic-core==2.27.2 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
pydantic==2.10.6 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
python-decouple==3.8 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
pytz==2025.1 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
segno==1.6.1 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
setuptools==75.8.0 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
sqlparse==0.5.3 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
tenacity==9.0.0 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
typing-extensions==4.12.2 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0"
tzdata==2025.1 ; python_full_version >= "3.11.6" and python_full_version < "4.0.0" and sys_platform == "win32"
